Implement card migration and enhance card model structure
- Introduced a new migration for card data, including the addition of card rarities and options.
- Updated the Card model to include new fields such as Code, RarityID, and Active status.
- Enhanced the UserInventory model to establish a foreign key relationship with the Card model.
- Refactored the DrawCard function to accommodate changes in card ID handling.
- Added new card codes for better identification and gameplay mechanics.

+++ services/cardService/card_registry.go
@@ -61,7 +61,7 @@ func PickRandomCard(hasSubscription bool) *models.Card {
 
 func GetCardByID(id int) *models.Card {
 	for i := range Deck {
-		if Deck[i].ID == id {
+		if Deck[i].ID == uint(id) {
 			return &Deck[i]
 		}
 	}
